refactor: :recycle: Separation of Concerns
