+++ fly_away/main.ts
@@ -1,3 +1,4 @@
+import * as flyMod from "./flightModule";
 import * as issMod from "./issModule";
 import * as mapMod from "./mapModule";
 
@@ -5,3 +6,4 @@ const map = mapMod.map;
 
 mapMod.drawMap(map);
 issMod.createISSMarker(map);
+flyMod.setFlights();
