+++ fly_away/src/models/IFlightAPIResponse.d.ts
@@ -1,0 +1,4 @@
+export interface IFLightAPIResponse {
+    time: string;
+    states: (string | number | boolean | number[])[][];
+}
