+++ fly_away/src/models/IFlightMarker.d.ts
@@ -1,0 +1,5 @@
+
+export interface IFlightMarker {
+    planeMarker: L.Marker;
+    plane: IFlight;
+}
