+++ fly_away/src/models/IISSResponse.d.ts
@@ -1,0 +1,15 @@
+export interface IISSAPIResponse {
+    name: string;
+    id: number;
+    latitude: number;
+    longitude: number;
+    altitude: number;
+    velocity: number;
+    visibility: string;
+    footprint: number;
+    timestamp: number;
+    daynum: number;
+    solar_lat: number;
+    solar_lon: number;
+    units: string;
+}
